refactor: 환불 계좌 정보 payment -> bookingCancel로 이동
